Support more devices
