+++ .gitignore
@@ -1,2 +1,1 @@
-__pycache__
 .idea
