--- const.py
@@ -1,1 +1,0 @@
-DOMAIN = "salus"
