+++ custom_components/salus/__init__.py
@@ -1,27 +1,23 @@
-"""The salus component."""
+"""Support for Salus iT600."""
 import logging
 
 from homeassistant import config_entries, core
-from homeassistant.const import CONF_HOST, CONF_TOKEN
-from homeassistant.helpers import device_registry as dr
 
 from .config_flow import CONF_FLOW_TYPE, CONF_USER
 from .const import DOMAIN
 
 _LOGGER = logging.getLogger(__name__)
 
-GATEWAY_PLATFORMS = ["climate", "binary_sensor"]
+GATEWAY_PLATFORMS = ["climate", "binary_sensor", "switch", "cover", "sensor"]
 
 
-async def async_setup(hass: core.HomeAssistant, config: dict):
-    """Set up the Salus component."""
+async def async_setup(hass: core.HomeAssistant, config: dict) -> bool:
+    """Set up the Salus iT600 component."""
     return True
 
 
-async def async_setup_entry(
-    hass: core.HomeAssistant, entry: config_entries.ConfigEntry
-):
-    """Set up the Salus components from a config entry."""
+async def async_setup_entry(hass: core.HomeAssistant, entry: config_entries.ConfigEntry):
+    """Set up components from a config entry."""
     hass.data[DOMAIN] = {}
     if entry.data[CONF_FLOW_TYPE] == CONF_USER:
         for component in GATEWAY_PLATFORMS:
