+++ custom_components/salus/const.py
@@ -1,0 +1,3 @@
+"""Constants of the Salus iT600 component."""
+
+DOMAIN = "salus"
