+++ custom_components/salus/strings.json
@@ -14,7 +14,7 @@
     },
     "error": {
       "connect_error": "Failed to connect, please check IP address",
-      "auth_error": "Failed to connect, please check token"
+      "auth_error": "Failed to connect, please check EUID"
     },
     "abort": {
       "already_configured": "Device is already configured",
