+++ custom_components/salus/translations/en.json
@@ -5,7 +5,7 @@
             "already_in_progress": "Config flow for this Salus device is already in progress."
         },
         "error": {
-            "auth_error": "Failed to connect, please check token",
+            "auth_error": "Failed to connect, please check EUID",
             "connect_error": "Failed to connect, please check IP address"
         },
         "flow_title": "Salus: {name}",
