+++ hacs.json
@@ -1,0 +1,5 @@
+{
+  "name": "Salus iT600",
+  "render_readme": true,
+  "iot_class": "Local Polling"
+}
